refactor: _src-like layout for mcmcloop
