--- src/bartz/mcmcloop.py
@@ -1,899 +1,0 @@
-# bartz/src/bartz/mcmcloop.py
-#
-# Copyright (c) 2024-2026, The Bartz Contributors
-#
-# This file is part of bartz.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-"""Functions that implement the full BART posterior MCMC loop.
-
-The entry points are `run_mcmc` and `make_default_callback`.
-"""
-
-import math
-from collections.abc import Callable
-from functools import partial, update_wrapper, wraps
-from typing import Any, NamedTuple, Protocol, TypeVar
-
-import jax
-import numpy
-from equinox import Module
-from jax import (
-    NamedSharding,
-    ShapeDtypeStruct,
-    debug,
-    device_put,
-    eval_shape,
-    jit,
-    lax,
-    named_call,
-    tree,
-    vmap,
-)
-from jax import numpy as jnp
-from jax.nn import softmax
-from jax.sharding import Mesh, PartitionSpec
-from jaxtyping import (
-    Array,
-    ArrayLike,
-    Bool,
-    Float32,
-    Int32,
-    Integer,
-    Key,
-    PyTree,
-    Shaped,
-    UInt,
-)
-from numpy.lib.array_utils import normalize_axis_index
-
-from bartz import mcmcstep
-from bartz._jaxext import autobatch, jit_active, split
-from bartz.grove import TreesTrace, evaluate_forest, forest_fill, var_histogram
-from bartz.mcmcstep import State
-from bartz.mcmcstep._state import (
-    CHAIN_AXIS,
-    add_dummy_axis,
-    chain_to_axis,
-    chain_vmap_axes,
-    chainful_axis,
-    field,
-    get_axis_size,
-    trace_sample_axes,
-)
-
-
-class BurninTrace(Module):
-    """MCMC trace with only diagnostic values."""
-
-    has_chains: bool = field(static=True)
-    """Whether the trace carries an explicit chain axis."""
-
-    error_cov_inv: (
-        Float32[Array, '*chains_and_samples']
-        | Float32[Array, '*chains_and_samples k k']
-    ) = field(chains=CHAIN_AXIS, samples=0)
-    theta: Float32[Array, '*chains_and_samples'] | None = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    grow_prop_count: Int32[Array, '*chains_and_samples'] = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    grow_acc_count: Int32[Array, '*chains_and_samples'] = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    prune_prop_count: Int32[Array, '*chains_and_samples'] = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    prune_acc_count: Int32[Array, '*chains_and_samples'] = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    log_likelihood: Float32[Array, '*chains_and_samples'] | None = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    log_trans_prior: Float32[Array, '*chains_and_samples'] | None = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-
-    @classmethod
-    def from_state(cls, state: State) -> 'BurninTrace':
-        """Create a single-item burn-in trace from a MCMC state."""
-        return cls(
-            has_chains=state.has_chains,
-            error_cov_inv=state.error_cov_inv,
-            theta=state.forest.theta,
-            grow_prop_count=state.forest.grow_prop_count,
-            grow_acc_count=state.forest.grow_acc_count,
-            prune_prop_count=state.forest.prune_prop_count,
-            prune_acc_count=state.forest.prune_acc_count,
-            log_likelihood=state.forest.log_likelihood,
-            log_trans_prior=state.forest.log_trans_prior,
-        )
-
-
-class MainTrace(BurninTrace):
-    """MCMC trace with trees and diagnostic values."""
-
-    leaf_tree: (
-        Float32[Array, '*chains_and_samples 2**d']
-        | Float32[Array, '*chains_and_samples k 2**d']
-    ) = field(chains=CHAIN_AXIS, samples=0)
-    var_tree: UInt[Array, '*chains_and_samples 2**(d-1)'] = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    split_tree: UInt[Array, '*chains_and_samples 2**(d-1)'] = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-    offset: Float32[Array, ''] | Float32[Array, ' k']
-    varprob: Float32[Array, '*chains_and_samples p'] | None = field(
-        chains=CHAIN_AXIS, samples=0
-    )
-
-    @classmethod
-    def from_state(cls, state: State) -> 'MainTrace':
-        """Create a single-item main trace from a MCMC state."""
-        # compute varprob
-        log_s = state.forest.log_s
-        if log_s is None:
-            varprob = None
-        else:
-            chain_axis = chain_vmap_axes(state.forest).log_s
-            p_axis = chainful_axis(0, chain_axis)  # (p,)
-            where = state.forest.max_split.astype(bool)
-            if chain_axis is not None:
-                where = jnp.expand_dims(where, chain_axis)
-            varprob = softmax(log_s, axis=p_axis, where=where)
-
-        return cls(
-            leaf_tree=state.forest.leaf_tree,
-            var_tree=state.forest.var_tree,
-            split_tree=state.forest.split_tree,
-            offset=state.offset,
-            varprob=varprob,
-            **vars(BurninTrace.from_state(state)),
-        )
-
-
-CallbackState = PyTree[Any, 'T']
-
-
-class RunMCMCResult(NamedTuple):
-    """Return value of `run_mcmc`."""
-
-    final_state: State
-    """The final MCMC state."""
-
-    burnin_trace: BurninTrace
-    """The trace of the burn-in phase."""
-
-    main_trace: MainTrace
-    """The trace of the main phase."""
-
-
-class Callback(Protocol):
-    """Callback type for `run_mcmc`."""
-
-    def __call__(
-        self,
-        *,
-        key: Key[Array, ''],
-        bart: State,
-        burnin: Bool[Array, ''],
-        i_total: Int32[Array, ''],
-        callback_state: CallbackState,
-        n_burn: Int32[Array, ''],
-        n_save: Int32[Array, ''],
-        n_skip: Int32[Array, ''],
-        i_outer: Int32[Array, ''],
-        inner_loop_length: Int32[Array, ''],
-    ) -> tuple[State, CallbackState] | None:
-        """Do an arbitrary action after an iteration of the MCMC.
-
-        Parameters
-        ----------
-        key
-            A key for random number generation.
-        bart
-            The MCMC state just after updating it.
-        burnin
-            Whether the last iteration was in the burn-in phase.
-        i_total
-            The index of the last MCMC iteration (0-based).
-        callback_state
-            The callback state, initially set to the argument passed to
-            `run_mcmc`, afterwards to the value returned by the last invocation
-            of the callback.
-        n_burn
-        n_save
-        n_skip
-            The corresponding `run_mcmc` arguments as-is.
-        i_outer
-            The index of the last outer loop iteration (0-based).
-        inner_loop_length
-            The number of MCMC iterations in the inner loop.
-
-        Returns
-        -------
-        bart : State
-            A possibly modified MCMC state. To avoid modifying the state,
-            return the `bart` argument passed to the callback as-is.
-        callback_state : CallbackState
-            The new state to be passed on the next callback invocation.
-
-        Notes
-        -----
-        For convenience, the callback may return `None`, and the states won't
-        be updated.
-        """
-        ...
-
-
-class _Carry(Module):
-    """Carry used in the loop in `run_mcmc`."""
-
-    bart: State
-    i_total: Int32[Array, '']
-    key: Key[Array, '']
-    burnin_trace: BurninTrace
-    main_trace: MainTrace
-    callback_state: CallbackState
-
-
-def run_mcmc(
-    key: Key[Array, ''],
-    bart: State,
-    n_save: int,
-    *,
-    n_burn: int = 0,
-    n_skip: int = 1,
-    inner_loop_length: int | None = None,
-    callback: Callback | None = None,
-    callback_state: CallbackState = None,
-) -> RunMCMCResult:
-    """
-    Run the MCMC for the BART posterior.
-
-    Parameters
-    ----------
-    key
-        A key for random number generation.
-    bart
-        The initial MCMC state, as created and updated by the functions in
-        `bartz.mcmcstep`. The MCMC loop uses buffer donation to avoid copies,
-        so this variable is invalidated after running `run_mcmc`. Make a copy
-        beforehand to use it again.
-    n_save
-        The number of iterations to save.
-    n_burn
-        The number of initial iterations which are not saved.
-    n_skip
-        The number of iterations to skip between each saved iteration, plus 1.
-        The effective burn-in is ``n_burn + n_skip - 1``.
-    inner_loop_length
-        The MCMC loop is split into an outer and an inner loop. The outer loop
-        is in Python, while the inner loop is in JAX. `inner_loop_length` is the
-        number of iterations of the inner loop to run for each iteration of the
-        outer loop. If not specified, the outer loop will iterate just once,
-        with all iterations done in a single inner loop run. The inner stride is
-        unrelated to the stride used for saving the trace.
-    callback
-        An arbitrary function run during the loop after updating the state. For
-        the signature, see `Callback`. The callback is called under the jax jit,
-        so the argument values are not available at the time the Python code is
-        executed. Use the utilities in `jax.debug` to access the values at
-        actual runtime. The callback may return new values for the MCMC state
-        and the callback state.
-    callback_state
-        The initial custom state for the callback.
-
-    Returns
-    -------
-    A namedtuple with the final state, the burn-in trace, and the main trace.
-
-    Raises
-    ------
-    RuntimeError
-        If `run_mcmc` detects it's being invoked in a `jit`-wrapped context and
-        with settings that would create unrolled loops in the trace.
-
-    Notes
-    -----
-    The number of MCMC updates is ``n_burn + n_skip * n_save``. The traces do
-    not include the initial state, and include the final state.
-    """
-    # create empty traces
-    burnin_trace = _empty_trace(n_burn, bart, BurninTrace)
-    main_trace = _empty_trace(n_save, bart, MainTrace)
-
-    # determine number of iterations for inner and outer loops
-    n_iters = n_burn + n_skip * n_save
-    if inner_loop_length is None:
-        inner_loop_length = n_iters
-    if inner_loop_length:
-        n_outer = n_iters // inner_loop_length + bool(n_iters % inner_loop_length)
-    else:
-        n_outer = 1
-        # setting to 0 would make for a clean noop, but it's useful to keep the
-        # same code path for benchmarking and testing
-
-    # error if under jit and there are unrolled loops
-    if jit_active() and n_outer > 1:
-        msg = (
-            '`run_mcmc` was called within a jit-compiled function and '
-            'there are more than 1 outer loops, '
-            'please either do not jit or set `inner_loop_length=None`'
-        )
-        raise RuntimeError(msg)
-
-    replicate = partial(_replicate, mesh=bart.config.mesh)
-    carry = _Carry(
-        bart,
-        replicate(jnp.int32(0)),
-        replicate(key),
-        burnin_trace,
-        main_trace,
-        callback_state,
-    )
-    _run_mcmc_inner_loop._fun.reset_call_counter()  # noqa: SLF001
-    for i_outer in range(n_outer):
-        carry = _run_mcmc_inner_loop(
-            carry, inner_loop_length, callback, n_burn, n_save, n_skip, i_outer, n_iters
-        )
-
-    return RunMCMCResult(carry.bart, carry.burnin_trace, carry.main_trace)
-
-
-def _replicate(x: Array, mesh: Mesh | None) -> Array:
-    if mesh is None:
-        return x
-    else:
-        return device_put(x, NamedSharding(mesh, PartitionSpec()))
-
-
-@partial(jit, static_argnums=(0, 2))
-def _empty_trace(length: int, bart: State, trace_cls: type[BurninTrace]) -> BurninTrace:
-    example_output = eval_shape(trace_cls.from_state, bart)
-    out_axes = trace_sample_axes(add_dummy_axis(example_output))
-
-    return vmap(
-        trace_cls.from_state, in_axes=None, out_axes=out_axes, axis_size=length
-    )(bart)
-
-
-T = TypeVar('T')
-
-
-class _CallCounter:
-    """Wrap a callable to check it's not called more than once."""
-
-    def __init__(self, func: Callable[..., T]) -> None:
-        self.func = func
-        self.n_calls = 0
-        update_wrapper(self, func)
-
-    def reset_call_counter(self) -> None:
-        """Reset the call counter."""
-        self.n_calls = 0
-
-    def __call__(self, *args: Any, **kwargs: Any) -> T:
-        if self.n_calls:
-            msg = (
-                'The inner loop of `run_mcmc` was traced more than once, '
-                'which indicates a double compilation of the MCMC code. This '
-                'probably depends on the input state having different type from the '
-                'output state. Check the input is in a format that is the '
-                'same jax would output, e.g., all arrays and scalars are jax '
-                'arrays, with the right shardings.'
-            )
-            raise RuntimeError(msg)
-        self.n_calls += 1
-        return self.func(*args, **kwargs)
-
-
-@partial(jit, donate_argnums=(0,), static_argnums=(2,))
-@_CallCounter
-def _run_mcmc_inner_loop(
-    carry: _Carry,
-    inner_loop_length: Int32[Array, ''],
-    callback: Callback | None,
-    n_burn: Int32[Array, ''],
-    n_save: Int32[Array, ''],
-    n_skip: Int32[Array, ''],
-    i_outer: Int32[Array, ''],
-    n_iters: Int32[Array, ''],
-) -> _Carry:
-    # determine number of iterations for this loop batch
-    i_upper = jnp.minimum(carry.i_total + inner_loop_length, n_iters)
-
-    def cond(carry: _Carry) -> Bool[Array, '']:
-        """Whether to continue the MCMC loop."""
-        return carry.i_total < i_upper
-
-    def body(carry: _Carry) -> _Carry:
-        """Update the MCMC state."""
-        # split random key
-        keys = split(carry.key, 3)
-        key = keys.pop()
-
-        # update state
-        bart = mcmcstep.step(keys.pop(), carry.bart)
-
-        # invoke callback
-        callback_state = carry.callback_state
-        if callback is not None:
-            rt = callback(
-                key=keys.pop(),
-                bart=bart,
-                burnin=carry.i_total < n_burn,
-                i_total=carry.i_total,
-                callback_state=callback_state,
-                n_burn=n_burn,
-                n_save=n_save,
-                n_skip=n_skip,
-                i_outer=i_outer,
-                inner_loop_length=inner_loop_length,
-            )
-            if rt is not None:
-                bart, callback_state = rt
-
-        # save to trace
-        burnin_trace, main_trace = _save_state_to_trace(
-            carry.burnin_trace, carry.main_trace, bart, carry.i_total, n_burn, n_skip
-        )
-
-        return _Carry(
-            bart=bart,
-            i_total=carry.i_total + 1,
-            key=key,
-            burnin_trace=burnin_trace,
-            main_trace=main_trace,
-            callback_state=callback_state,
-        )
-
-    return lax.while_loop(cond, body, carry)
-
-
-@named_call
-def _save_state_to_trace(
-    burnin_trace: BurninTrace,
-    main_trace: MainTrace,
-    bart: State,
-    i_total: Int32[Array, ''],
-    n_burn: Int32[Array, ''],
-    n_skip: Int32[Array, ''],
-) -> tuple[BurninTrace, MainTrace]:
-    # trace index where to save during burnin; out-of-bounds => noop after
-    # burnin
-    burnin_idx = i_total
-
-    # trace index where to save during main phase; force it out-of-bounds
-    # during burnin
-    main_idx = (i_total - n_burn) // n_skip
-    noop_idx = jnp.iinfo(jnp.int32).max
-    noop_cond = i_total < n_burn
-    main_idx = jnp.where(noop_cond, noop_idx, main_idx)
-
-    # prepare array index
-    burnin_trace = _set(burnin_trace, burnin_idx, BurninTrace.from_state(bart))
-    main_trace = _set(main_trace, main_idx, MainTrace.from_state(bart))
-
-    return burnin_trace, main_trace
-
-
-def _set(
-    trace: PyTree[Array, ' T'], index: Int32[Array, ''], val: PyTree[Array, ' T']
-) -> PyTree[Array, ' T']:
-    """Do ``trace[index] = val`` but fancier."""
-    sample_axes = trace_sample_axes(trace)
-
-    def at_set(
-        trace: Shaped[Array, 'chains samples *shape']
-        | Shaped[Array, ' samples *shape'],
-        val: Shaped[Array, ' chains *shape'] | Shaped[Array, '*shape'],
-        sample_axis: int | None,
-    ) -> Shaped[Array, 'chains samples *shape']:
-        if sample_axis is None or trace.size == 0:
-            # `sample_axis is None`: fields without a `samples` marker have
-            # no per-iteration slot to update.
-            # `trace.size == 0`: jax refuses to index into an axis of length
-            # 0, even in the abstract.
-            return trace
-
-        ndindex = (slice(None),) * sample_axis + (index, ...)
-        return trace.at[ndindex].set(val, mode='drop')
-
-    return tree.map(at_set, trace, val, sample_axes)
-
-
-def make_default_callback(
-    state: State,
-    *,
-    dot_every: int | Integer[Array, ''] | None = 1,
-    report_every: int | Integer[Array, ''] | None = 100,
-) -> dict[str, Any]:
-    """
-    Prepare a default callback for `run_mcmc`.
-
-    The callback prints a dot on every iteration, and a longer
-    report outer loop iteration, and can do variable selection.
-
-    Parameters
-    ----------
-    state
-        The bart state to use the callback with, used to determine device
-        sharding.
-    dot_every
-        A dot is printed every `dot_every` MCMC iterations, `None` to disable.
-    report_every
-        A one line report is printed every `report_every` MCMC iterations,
-        `None` to disable.
-
-    Returns
-    -------
-    A dictionary with the arguments to pass to `run_mcmc` as keyword arguments to set up the callback.
-
-    Examples
-    --------
-    >>> run_mcmc(key, state, ..., **make_default_callback(state, ...))
-    """
-
-    def as_replicated_array_or_none(val: ArrayLike | None) -> None | Array:
-        return None if val is None else _replicate(jnp.asarray(val), state.config.mesh)
-
-    return dict(
-        callback=print_callback,
-        callback_state=PrintCallbackState(
-            as_replicated_array_or_none(dot_every),
-            as_replicated_array_or_none(report_every),
-        ),
-    )
-
-
-class PrintCallbackState(Module):
-    """State for `print_callback`."""
-
-    dot_every: Int32[Array, ''] | None
-    """A dot is printed every `dot_every` MCMC iterations, `None` to disable."""
-
-    report_every: Int32[Array, ''] | None
-    """A one line report is printed every `report_every` MCMC iterations,
-    `None` to disable."""
-
-
-def print_callback(
-    *,
-    bart: State,
-    burnin: Bool[Array, ''],
-    i_total: Int32[Array, ''],
-    n_burn: Int32[Array, ''],
-    n_save: Int32[Array, ''],
-    n_skip: Int32[Array, ''],
-    callback_state: PrintCallbackState,
-    **_: Any,
-) -> None:
-    """Print a dot and/or a report periodically during the MCMC."""
-    report_every = callback_state.report_every
-    dot_every = callback_state.dot_every
-    it = i_total + 1
-
-    def get_cond(every: Int32[Array, ''] | None) -> bool | Bool[Array, '']:
-        return False if every is None else it % every == 0
-
-    report_cond = get_cond(report_every)
-    dot_cond = get_cond(dot_every)
-
-    def line_report_branch() -> None:
-        if report_every is None:
-            return
-        if dot_every is None:
-            print_newline = False
-        else:
-            print_newline = it % report_every > it % dot_every
-        chain_axis = chain_vmap_axes(bart.forest).split_tree
-        num_trees_axis = chainful_axis(0, chain_axis)  # (num_trees, hts)
-        split_tree = chain_to_axis(bart.forest.split_tree, chain_axis)
-        debug.callback(
-            _print_report,
-            print_dot=dot_cond,
-            print_newline=print_newline,
-            burnin=burnin,
-            it=it,
-            n_iters=n_burn + n_save * n_skip,
-            num_chains=bart.num_chains(),
-            grow_prop_count=bart.forest.grow_prop_count.mean(),
-            grow_acc_count=bart.forest.grow_acc_count.mean(),
-            prune_acc_count=bart.forest.prune_acc_count.mean(),
-            prop_total=bart.forest.split_tree.shape[num_trees_axis],
-            fill=forest_fill(split_tree),
-        )
-
-    def just_dot_branch() -> None:
-        if dot_every is None:
-            return
-        debug.callback(
-            lambda: print('.', end='', flush=True)  # noqa: T201
-        )
-        # logging can't do in-line printing so we use print
-
-    lax.cond(
-        report_cond,
-        line_report_branch,
-        lambda: lax.cond(dot_cond, just_dot_branch, lambda: None),
-    )
-
-
-def _convert_jax_arrays_in_args(func: Callable[..., T]) -> Callable[..., T]:
-    """Remove jax arrays from a function arguments.
-
-    Converts all `jax.Array` instances in the arguments to either Python scalars
-    or numpy arrays.
-    """
-
-    def convert_jax_arrays(pytree: PyTree) -> PyTree:
-        def convert_jax_array(val: object) -> object:
-            if not isinstance(val, Array):
-                return val
-            elif val.shape:
-                return numpy.array(val)
-            else:
-                return val.item()
-
-        return tree.map(convert_jax_array, pytree)
-
-    @wraps(func)
-    def new_func(*args: Any, **kw: Any) -> T:
-        args = convert_jax_arrays(args)
-        kw = convert_jax_arrays(kw)
-        return func(*args, **kw)
-
-    return new_func
-
-
-@_convert_jax_arrays_in_args
-# convert all jax arrays in arguments because operations on them could lead to
-# deadlock with the main thread
-def _print_report(
-    *,
-    print_dot: bool,
-    print_newline: bool,
-    burnin: bool,
-    it: int,
-    n_iters: int,
-    num_chains: int | None,
-    grow_prop_count: float,
-    grow_acc_count: float,
-    prune_acc_count: float,
-    prop_total: int,
-    fill: float,
-) -> None:
-    """Print the report for `print_callback`."""
-    # compute fractions
-    grow_prop = grow_prop_count / prop_total
-    move_acc = (grow_acc_count + prune_acc_count) / prop_total
-
-    # determine prefix
-    if print_dot:
-        prefix = '.\n'
-    elif print_newline:
-        prefix = '\n'
-    else:
-        prefix = ''
-
-    # determine suffix in parentheses
-    msgs = []
-    if num_chains is not None:
-        msgs.append(f'avg. {num_chains} chains')
-    if burnin:
-        msgs.append('burnin')
-    suffix = f' ({", ".join(msgs)})' if msgs else ''
-
-    print(  # noqa: T201, see print_callback for why not logging
-        f'{prefix}Iteration {it}/{n_iters}, '
-        f'grow prob: {grow_prop:.0%}, '
-        f'move acc: {move_acc:.0%}, '
-        f'fill: {fill:.0%}{suffix}'
-    )
-
-
-@partial(jit, static_argnames=('flatten_chains', 'out_chain_axis_w_trees'))
-def evaluate_trace(
-    X: UInt[Array, 'p n'],
-    trace: MainTrace,
-    *,
-    flatten_chains: bool = False,
-    out_chain_axis_w_trees: int = CHAIN_AXIS,
-) -> Float32[Array, '*trace_shape n'] | Float32[Array, '*trace_shape k n']:
-    """
-    Compute predictions for all iterations of the BART MCMC.
-
-    Parameters
-    ----------
-    X
-        The predictors matrix, with `p` predictors and `n` observations.
-    trace
-        A main trace of the BART MCMC, as returned by `run_mcmc`.
-    flatten_chains
-        If `True` and `trace` has a chain axis, collapse it into the sample
-        axis, so the leading dimension of the output is ``num_chains *
-        num_samples`` instead of ``(num_chains, num_samples)``.
-    out_chain_axis_w_trees
-        Position of the chain axis in the output. Interpreted against the
-        intermediate, pre-tree-reduction layout ``(sample, tree, *k, n)``;
-        after summing over the tree axis the chain ends up one position
-        earlier if it was after the tree axis. Negative values count from the
-        end. Ignored when `trace` has no chain axis or `flatten_chains` is
-        True.
-
-    Returns
-    -------
-    The predictions for each chain and iteration of the MCMC.
-    """
-    # per-device memory limit
-    max_io_nbytes = 2**27  # 128 MiB
-
-    # adjust memory limit for number of devices
-    mesh = jax.typeof(trace.leaf_tree).sharding.mesh
-    num_devices = get_axis_size(mesh, 'chains') * get_axis_size(mesh, 'data')
-    max_io_nbytes *= num_devices
-
-    # extract only the trees from the trace, this will be the input to `evaluate_forest`
-    trees = TreesTrace.from_dataclass(trace)
-    batched_eval = evaluate_forest  # we will transform `batched_eval`
-
-    # determine batching axes
-    trace_chain_axes = chain_vmap_axes(trace)
-    # WORKAROUND(python<3.14): use operator.is_none
-    is_none = lambda x: x is None
-    sample_axes = tree.map(partial(chainful_axis, 0), trace_chain_axes, is_leaf=is_none)
-    tree_axes = tree.map(partial(chainful_axis, 1), trace_chain_axes, is_leaf=is_none)
-
-    # stuff to determine output shapes
-    # leaf_tree has shape (sample, tree, *k, ts)
-    k_axis = chainful_axis(2, trace_chain_axes.leaf_tree)
-    is_mv = trace.leaf_tree.ndim > trace.split_tree.ndim
-    kshape = trace.leaf_tree.shape[k_axis : k_axis + is_mv]
-    _, n = X.shape
-    num_samples = trace.leaf_tree.shape[sample_axes.leaf_tree]
-    num_trees = trace.leaf_tree.shape[tree_axes.leaf_tree]
-    if trace.has_chains:
-        num_chains = trace.leaf_tree.shape[trace_chain_axes.leaf_tree]
-
-    def expand_shape(shape: tuple[int, ...], chain_axis: int) -> tuple[int, ...]:
-        return (*shape[:chain_axis], num_chains, *shape[chain_axis:])
-
-    # tree axis to reduce over
-    out_shape_w_trees = (num_samples, num_trees, *kshape, n)
-    tree_axis = 1
-    if trace.has_chains:
-        out_chain_axis_w_trees = normalize_axis_index(
-            out_chain_axis_w_trees, 1 + len(out_shape_w_trees)
-        )
-        tree_axis = chainful_axis(tree_axis, out_chain_axis_w_trees)
-
-    # vmap over chains
-    if trace.has_chains:
-        batched_eval = vmap(
-            batched_eval,
-            in_axes=(None, TreesTrace.from_dataclass(trace_chain_axes)),
-            out_axes=out_chain_axis_w_trees,
-        )
-
-    # batch and sum over trees
-    batched_eval = autobatch(
-        batched_eval,
-        max_io_nbytes,
-        in_axes=(None, TreesTrace.from_dataclass(tree_axes)),
-        out_axes=tree_axis,
-        reduce_ufunc=jnp.add,
-    )
-
-    # output shape after reducing trees
-    out_shape = (num_samples, *kshape, n)
-    sample_axis = 0
-    if trace.has_chains:
-        out_chain_axis = out_chain_axis_w_trees - (out_chain_axis_w_trees > tree_axis)
-        sample_axis = chainful_axis(0, out_chain_axis)
-        out_shape = expand_shape(out_shape, out_chain_axis)
-
-    # adjust memory limit keeping into account that trees are summed over
-    # split_tree has shape (sample, tree, hts)
-    hts = trace.split_tree.shape[chainful_axis(2, trace_chain_axes.split_tree)]
-    k = math.prod(kshape)
-    out_size = k * n * jnp.float32.dtype.itemsize  # the value of the forest
-    core_io_size = (
-        num_trees
-        * hts
-        * (
-            2 * k * trace.leaf_tree.itemsize
-            + trace.var_tree.itemsize
-            + trace.split_tree.itemsize
-        )
-        + out_size
-    )
-    core_int_size = (num_trees - 1) * out_size
-    max_io_nbytes = max(
-        1, math.floor(max_io_nbytes / (1 + core_int_size / core_io_size))
-    )
-
-    # batch over mcmc samples
-    batched_eval = autobatch(
-        batched_eval,
-        max_io_nbytes,
-        in_axes=(None, TreesTrace.from_dataclass(sample_axes)),
-        out_axes=sample_axis,
-        warn_on_overflow=False,  # the inner autobatch will handle it
-        result_shape_dtype=ShapeDtypeStruct(out_shape, jnp.float32),
-    )
-
-    # prepare offset for broadcasting
-    # chainless y shape = (num_samples, *kshape, n), offset shape = kshape
-    offset = trace.offset[None, ..., None]
-    if trace.has_chains:
-        offset = jnp.expand_dims(offset, out_chain_axis)
-
-    # evaluate trees
-    y_centered = batched_eval(X, trees)
-    y = y_centered + offset
-    if flatten_chains and trace.has_chains:
-        y = jnp.moveaxis(y, out_chain_axis, 0)
-        y = lax.collapse(y, 0, 2)
-    return y
-
-
-@partial(jit, static_argnames=('p', 'out_chain_axis'))
-def compute_varcount(
-    p: int, trace: MainTrace, *, out_chain_axis: int = CHAIN_AXIS
-) -> Int32[Array, '*trace_shape {p}']:
-    """
-    Count how many times each predictor is used in each MCMC state.
-
-    Parameters
-    ----------
-    p
-        The number of predictors.
-    trace
-        A main trace of the BART MCMC, as returned by `run_mcmc`.
-    out_chain_axis
-        Position of the chain axis in the output, whose chainless shape is
-        ``(samples, p)``. Negative values count from the end. Ignored when
-        `trace` has no chain axis.
-
-    Returns
-    -------
-    Histogram of predictor usage in each MCMC state.
-    """
-    chain_axes = chain_vmap_axes(trace)
-
-    def histogram(
-        var_tree: UInt[Array, 'samples trees nodes'],
-        split_tree: UInt[Array, 'samples trees nodes'],
-    ) -> Int32[Array, 'samples {p}']:
-        return var_histogram(p, var_tree, split_tree, sum_batch_axis=-1)
-
-    if trace.has_chains:
-        # chainless output ndim is 2 (samples, p); the chain adds one
-        out_axis = normalize_axis_index(out_chain_axis, 3)
-        histogram = vmap(
-            histogram,
-            in_axes=(chain_axes.var_tree, chain_axes.split_tree),
-            out_axes=out_axis,
-        )
-
-    return histogram(trace.var_tree, trace.split_tree)
