+++ src/bartz/mcmcloop/__init__.py
@@ -1,0 +1,39 @@
+# bartz/src/bartz/mcmcloop/__init__.py
+#
+# Copyright (c) 2024-2026, The Bartz Contributors
+#
+# This file is part of bartz.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+"""Functions that implement the full BART posterior MCMC loop.
+
+The entry points are `run_mcmc` and `make_default_callback`.
+"""
+
+# ruff: noqa: F401
+
+from bartz.mcmcloop._callback import (
+    PrintCallbackState,
+    make_default_callback,
+    print_callback,
+)
+from bartz.mcmcloop._evaluate import compute_varcount, evaluate_trace
+from bartz.mcmcloop._loop import Callback, CallbackState, RunMCMCResult, run_mcmc
+from bartz.mcmcloop._trace import BurninTrace, MainTrace
